User-Characters relationships finished

+++ Praksa/services/CharacterService/CharacterService.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 using Praksa.Data;
 using Praksa.Dtos.Character;
@@ -6,6 +7,7 @@ using Praksa.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 
 namespace Praksa.services.CharacterService
@@ -15,20 +17,28 @@ namespace Praksa.services.CharacterService
         
         private readonly IMapper _mapper;
         private readonly DataContext _context;
+        private readonly IHttpContextAccessor _httpContextAccessor;
 
-        public CharacterService(IMapper mapper, DataContext context)
+        public CharacterService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor)
         {
             _mapper = mapper;
             _context = context;
+            _httpContextAccessor = httpContextAccessor;
         }
+
+        private int GetUserId() => int.Parse(_httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
 
         public async Task<ServiceResponse<List<GetCharacterDto>>> AddCharacter(AddCharacterDto newCharacter)
             {
                 var ServiceResponse = new ServiceResponse<List<GetCharacterDto>>();
                 Character character = _mapper.Map<Character>(newCharacter);
+                character.User = await _context.Users.FirstOrDefaultAsync(u => u.Id == GetUserId());
+
                 _context.Characters.Add(character);
                 await _context.SaveChangesAsync();
-                ServiceResponse.Data = await _context.Characters.Select(c=> _mapper.Map<GetCharacterDto>(c)).ToListAsync();
+                ServiceResponse.Data = await _context.Characters
+                .Where(c => c.User.Id == GetUserId())
+                .Select(c=> _mapper.Map<GetCharacterDto>(c)).ToListAsync();
                 return ServiceResponse;
             }
 
@@ -37,13 +47,23 @@ namespace Praksa.services.CharacterService
             var ServiceResponse = new ServiceResponse<List<GetCharacterDto>>();
             try
             {
-                Character character = await _context.Characters.FirstAsync(c => c.Id == id);
+                Character character = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
 
-                _context.Characters.Remove(character);
+                if (character != null)
+                {
+                    _context.Characters.Remove(character);
 
-                await _context.SaveChangesAsync();
+                    await _context.SaveChangesAsync();
 
-                ServiceResponse.Data = _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+                    ServiceResponse.Data = _context.Characters
+                        .Where(_c => _c.User.Id == GetUserId())
+                        .Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+                } else
+                {
+                    ServiceResponse.Success = false;
+                    ServiceResponse.message = "Character not found.";
+                }
+                return ServiceResponse;
             }
             catch (Exception ex)
             {
@@ -53,10 +73,10 @@ namespace Praksa.services.CharacterService
             return ServiceResponse;
         }
 
-        public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters(int userId)
+        public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters()
             {
                 var ServiceResponse = new ServiceResponse<List<GetCharacterDto>>();
-                var dbCharacters = await _context.Characters.Where(c => c.User.Id == userId).ToListAsync();
+                var dbCharacters = await _context.Characters.Where(c => c.User.Id == GetUserId()).ToListAsync();
                 ServiceResponse.Data = dbCharacters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
                 return ServiceResponse;
             }
@@ -64,7 +84,7 @@ namespace Praksa.services.CharacterService
             public async Task<ServiceResponse<GetCharacterDto>> GetCharacterById(int id)
             {
                 var ServiceResponse = new ServiceResponse<GetCharacterDto>();
-            var dbCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id);
+            var dbCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
                 ServiceResponse.Data = _mapper.Map<GetCharacterDto>(dbCharacter);
                 return ServiceResponse;
             }
